Move JavaClassUtil to :messaging
This is the only production code that uses it.
