+++ platforms/jvm/language-java/src/integTest/groovy/org/gradle/api/tasks/compile/CrossCompilationIntegrationTest.groovy
@@ -25,7 +25,7 @@ import org.gradle.test.fixtures.archive.JarTestFixture
 import org.gradle.util.GradleVersion
 import org.junit.Assume
 
-import static org.gradle.internal.classanalysis.JavaClassUtil.getClassMajorVersion
+import static org.gradle.internal.serialize.JavaClassUtil.getClassMajorVersion
 
 @Flaky(because = "https://github.com/gradle/gradle-private/issues/3901")
 class CrossCompilationIntegrationTest extends AbstractIntegrationSpec {
