+++ platforms/jvm/plugins-java/build.gradle.kts
@@ -34,6 +34,7 @@ dependencies {
 
     testImplementation(testFixtures(project(":core")))
 
+    integTestImplementation(testFixtures(projects.messaging))
     integTestImplementation(testFixtures(project(":enterprise-operations")))
     integTestImplementation(testFixtures(project(":language-java")))
     integTestImplementation(testFixtures(project(":language-jvm")))
