+++ testing/architecture-test/src/changes/archunit_store/internal-api-nullability.txt
@@ -1758,7 +1758,6 @@ Class <org.gradle.internal.buildoption.StringBuildOption> is not annotated (dire
 Class <org.gradle.internal.buildoption.StringInternalOption> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (StringInternalOption.java:0)
 Class <org.gradle.internal.cache.MonitoredCleanupAction> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (MonitoredCleanupAction.java:0)
 Class <org.gradle.internal.classanalysis.AsmConstants> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (AsmConstants.java:0)
-Class <org.gradle.internal.classanalysis.JavaClassUtil> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JavaClassUtil.java:0)
 Class <org.gradle.internal.classpath.CachedClasspathTransformer$StandardTransform> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (CachedClasspathTransformer.java:0)
 Class <org.gradle.internal.classpath.CachedClasspathTransformer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (CachedClasspathTransformer.java:0)
 Class <org.gradle.internal.classpath.ClassData> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ClassData.java:0)
@@ -2762,6 +2761,7 @@ Class <org.gradle.internal.serialize.HierarchicalNameSerializer> is not annotate
 Class <org.gradle.internal.serialize.InputStreamBackedDecoder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (InputStreamBackedDecoder.java:0)
 Class <org.gradle.internal.serialize.IntSetSerializer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (IntSetSerializer.java:0)
 Class <org.gradle.internal.serialize.InterningStringSerializer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (InterningStringSerializer.java:0)
+Class <org.gradle.internal.serialize.JavaClassUtil> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JavaClassUtil.java:0)
 Class <org.gradle.internal.serialize.ListSerializer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ListSerializer.java:0)
 Class <org.gradle.internal.serialize.MapSerializer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (MapSerializer.java:0)
 Class <org.gradle.internal.serialize.Message> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (Message.java:0)
